% Generated by build script
\name{fda_drug_ndc}
\alias{fda_drug_ndc}
\title{Search the National Drug Code (NDC) directory}
\usage{
fda_drug_ndc()
}
\arguments{
\item{search}{Search query. Examples:}
\item{api_key}{Optional API key}
\item{max_results}{Max results (default 100)}
}
\value{tibble: product_ndc, brand_name, generic_name, dosage_form,}
\description{Search the National Drug Code (NDC) directory}
